refactor(ensindexer): rename realtime indexing status files
